Cleanup and restructuring for V1.
